refactor cpu attn
